Move chart packaging to the build
This was a bigger change than you'd think.

Instead of building the image, then running the app build
in that image's container,
We do all the app building in the docker build.

This means if we want app build logs separated from docker build,
we have to create a container based on the image and copy the files out.